+++ src/parser/monster/templates/monster.js
@@ -5,10 +5,12 @@ export async function newNPC(name) {
   };
   const npcClass = await Actor.create({ name, type: "npc" }, options);
   let npc = npcClass.toObject();
+  npc._id = randomID();
   const flags = {
     dnd5e: {},
     monsterMunch: {},
     ddbimporter: {
+      compendiumId: npc._id,
       dndbeyond: {},
     },
   };