+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # NEXT UP
 
 - Allow Avatar and Token swapping to both be enabled.
+- Chris Premades Support: Sneak Attack would not match on Monsters.
+- Chris Premades/Monster Item import: When monsters are moved from the DDB Compendium to the world, DDB Importer will attempt to fix the origins of effects it has generated, if possible.
+
 
 # 3.5.3